WIP - security blueprint & logout view
